pyproject.toml
[project]
name = "mpc-projects-login"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydoll-python>=2.8.0",
]
